+++ RELEASENOTES.md
@@ -2,6 +2,14 @@
 
 ### dev-v2 (not yet released) ###
 
+* Add `allowedCapturePolicy` field to `AudioAttributes` wrapper to allow to
+  opt-out of audio recording.
+* Add `DataSpec.httpRequestHeaders` to set HTTP request headers when connecting
+  to an HTTP source. `DefaultHttpDataSource`, `CronetDataSource` and
+  `OkHttpDataSource` include headers set in the DataSpec when connecting to the
+  source.
+* Bypass sniffing in `ProgressiveMediaPeriod` in case a single extractor is
+  provided ([#6325](https://github.com/google/ExoPlayer/issues/6325)).
 * Surface information provided by methods `isHardwareAccelerated`,
   `isSoftwareOnly` and `isVendor` added in Android Q in `MediaCodecInfo` class
   ([#5839](https://github.com/google/ExoPlayer/issues/5839)).
@@ -51,11 +59,19 @@
   ([#6335](https://github.com/google/ExoPlayer/issues/6335)).
 * Add `HttpDataSource.getResponseCode` to provide the status code associated
   with the most recent HTTP response.
-* Upgrade LibRtmp-Client-for-Android to fix RTMP playback issues
+* OkHttp extension: Upgrade OkHttp to fix HTTP2 socket timeout issue
+  ([#4078](https://github.com/google/ExoPlayer/issues/4078)).
+* RTMP extension: Upgrade LibRtmp-Client-for-Android to fix RTMP playback issues
   ([#4200](https://github.com/google/ExoPlayer/issues/4200),
   [#4249](https://github.com/google/ExoPlayer/issues/4249),
   [#4319](https://github.com/google/ExoPlayer/issues/4319),
   [#4337](https://github.com/google/ExoPlayer/issues/4337)).
+* Publish `testutils` module to simplify unit testing with ExoPlayer
+  ([#6267](https://github.com/google/ExoPlayer/issues/6267)).
+* Add `uid` to `Timeline.Window` to uniquely identify window instances.
+* Fix decoder selection for E-AC3 JOC streams
+  ([#6398](https://github.com/google/ExoPlayer/issues/6398)).
+* Fix Dolby Vision fallback to AVC and HEVC.
 
 ### 2.10.4 ###
 
@@ -87,6 +103,8 @@
 * Fix issue where playback speeds are not used in adaptive track selections
   after manual selection changes for other renderers
   ([#6256](https://github.com/google/ExoPlayer/issues/6256)).
+* Fix initialization data handling for FLAC in MP4
+  ([#6396](https://github.com/google/ExoPlayer/issues/6396)).
 
 ### 2.10.3 ###